requirements.txt
# - Deep Learning
torchvision
torchinfo
 
# - Explainability
grad-cam
 
# - Experiment Tracking
mlflow

# - Data versioning
dvc
 
# - Computer Vision 
opencv-python
Pillow
 
# - Data & Analisis
numpy
pandas
scikit-learn
split-folders
 
# - Visualisasi 
matplotlib
seaborn
 
# - Notebook & Utilities
# jupyter
# ipykernel
tqdm